+++ src/core/Portfolio/IPortfolioStorage.cs
