+++ src/core/Shared/HandlerWithStorage.cs
