+++ src/core/Shared/TickerPrice.cs
