+++ src/core/Stocks/Handlers/Dashboard.cs
