+++ src/core/Stocks/Handlers/Delete.cs
