+++ src/core/Stocks/Handlers/DeleteTransaction.cs
