+++ src/core/Stocks/Handlers/Details.cs
