+++ src/core/Stocks/Handlers/Export.cs
