+++ src/core/Stocks/Handlers/Get.cs
