+++ src/core/Stocks/Handlers/Search.cs
