+++ src/core/Stocks/Handlers/StockLists.cs
