--- src/core/Stocks/IStockTransaction.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace core.Stocks
-{
-    internal interface IStockTransaction
-    {
-        int NumberOfShares { get; }
-        double Price { get; }
-        Guid Id { get; }
-        DateTimeOffset When { get; }
-    }
-}
