--- src/core/Stocks/StockDeleted.cs
@@ -1,12 +1,0 @@
-using System;
-using core.Shared;
-
-namespace core.Stocks
-{
-    internal class StockDeleted : AggregateEvent
-    {
-        public StockDeleted(Guid id, Guid aggregateId, DateTimeOffset when) : base(id, aggregateId, when)
-        {
-        }
-    }
-}
