+++ src/core/Stocks/View/StockOwnershipPerformance.cs
