+++ src/core/Stocks/View/StockStatsContainer.cs
